% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csuv_plot.r
\name{csuv.plot.helper}
\alias{csuv.plot.helper}
\title{Helper function, please do not use it}
\usage{
csuv.plot.helper(
  new.fit,
  with.unconditional = FALSE,
  compare.method.fit = NULL,
  compare.method.names = NULL,
  cv.mod = NULL,
  print.compare.method.points = FALSE,
  ci.method = "conditional",
  with.thr = TRUE,
  with.violin = FALSE,
  to.shade = TRUE,
  level = 0.1,
  var.freq.thr = 0.1,
  ...
)
}
\arguments{
\item{new.fit}{fitted results from CSUV::csuv()}

\item{with.unconditional}{TRUE to get a unconditonal boxplot on the same graph. Default is FALSE}

\item{compare.method.fit}{(optional) fitted results from CSUV::lm.compare.methods()}

\item{compare.method.names}{(optional) names of method to compare}

\item{cv.mod}{(optional) fitted results from cross validation}

\item{print.compare.method.points}{Default is FALSE}

\item{ci.method}{how the confidence interval should be calculated. Default is "conditional"}

\item{with.thr}{whether the selection by the CSUV should be show. Default is TRUE}

\item{with.violin}{whether the graph with violin plot}

\item{to.shade}{whether to shade the graph by the relative frequency calculated by CSUV. Default is TRUE}

\item{level}{the significant level of the whiskers. Default is 0.1}

\item{var.freq.thr}{minimum variable frequency to show, default is 0.1}

\item{...}{additional argument for plot}
}
\value{
a ggplot object
}
\description{
Helper function, please do not use it
}
