% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_method.r
\name{get.compare.fit}
\alias{get.compare.fit}
\title{Helper function, please do not use}
\usage{
get.compare.fit(x, y, intercept, method.names, current.compare.fit = NULL)
}
\arguments{
\item{x}{covariates (n times p matrix, n: number of entries, p: number of covariates)}

\item{y}{response (vector with n entries)}

\item{intercept}{TRUE to fit the data with an intercept, FALSE to fit the data without an intercept}

\item{method.names}{vector of method names to be used in cross validation. Choose among "lasso", "elastic", "relaxo", "mcp" and "scad". Default is to use all methods listed above}

\item{current.compare.fit}{(optional)}
}
\value{
a list which includes the estimated coefficients (est.b) and the corresponding ordinary least square fit from stats::lm()
}
\description{
Helper function, please do not use
}
