% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_method.r
\name{get.compare.methods}
\alias{get.compare.methods}
\title{Get a list of variable selection methods implemented in the CSUV package}
\usage{
get.compare.methods()
}
\value{
a list of functions
}
\description{
Get a list of variable selection methods implemented in the CSUV package
}
\examples{
X = matrix(rnorm(1000), nrow = 100)
Y = rowSums(X[,1:3])+rnorm(100)
lasso.method = get.compare.methods()$lasso
lasso.mod = lasso.method(X, Y, intercept = FALSE)
print(lasso.mod$est.b)
}
