% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csuv.r
\name{get.csuv.final.mod}
\alias{get.csuv.final.mod}
\title{Helper function, please do not use it}
\usage{
get.csuv.final.mod(
  X,
  Y,
  intercept,
  unique.fit,
  selection.criterion,
  coef.est.method = lm.ols,
  q,
  method.names,
  B
)
}
\arguments{
\item{X}{covariates (n times p matrix, n: number of entries, p: number of covariates)}

\item{Y}{response (vector with n entries)}

\item{intercept}{TRUE to fit the data with an intercept, FALSE to fit the data without an intercept}

\item{unique.fit}{from get.csuv.unique.fit}

\item{selection.criterion}{= c("mse", "ebic"). Measure to select fitted models in subsampling dataset. "mse" is mean square error and "ebic" is extended BIC. Default is mse}

\item{coef.est.method}{method to estimate the coefficients of covariates after variable selection. User can provide his/her function. Default is ordinary least square}

\item{q}{percentile of fitted models used per each subsampling in CSUV, according to the selection criterion on out-of-sample data in ascending order. Default is q = 0 (only the fitted model with the lowest MSE in a subsampling data is used)}

\item{method.names}{vector of method names to be used in CSUV. Choose among "lasso", "elastic", "relaxo", "mcp" and "scad". Default is to use all methods listed above}

\item{B}{number of subsampling. Default is 100}
}
\value{
a list of current fit
}
\description{
Helper function, please do not use it
}
