% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{interactive.uncertainty.illustration}
\alias{interactive.uncertainty.illustration}
\title{Interactive version of the uncertainty illustration}
\usage{
interactive.uncertainty.illustration()
}
\description{
Interactive version of the uncertainty illustration
}
\examples{
\dontrun{
interactive.uncertainty.illustration()
}
}
