% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csuv_plot.r
\name{plot.csuv}
\alias{plot.csuv}
\title{Graphical illustration of selection uncertainty}
\usage{
\method{plot}{csuv}(x, is.box.plot = TRUE, compare.method.fit = NULL,
  cv.mod = NULL, with.thr = TRUE, to.shade = TRUE,
  ci.type = "original", level = 0.1, ylim = NULL, xlim = NULL,
  var.freq.thr = 0.05, return.lim = FALSE, ...)
}
\arguments{
\item{x}{fitted results from CSUV::csuv()}

\item{is.box.plot}{TRUE to get a box plot like uncertainty illustration. FALSE to get a "confidence interval like" uncertainty illustration. Default is TRUE}

\item{compare.method.fit}{(optional) fitted results from CSUV::lm.compare.methods()}

\item{cv.mod}{(optional) fitted results from cross validation}

\item{with.thr}{whether the selection by the CSUV should be show. Default is TRUE}

\item{to.shade}{whether to shade the graph by the relative frequency calculated by CSUV. Default is TRUE}

\item{ci.type}{todo}

\item{level}{(only use when is.box.plot is FALSE) the significant level of the "confidence interval". Default is 0.1}

\item{ylim}{range of y shown in the graph}

\item{xlim}{range of x shown in the graph}

\item{var.freq.thr}{minimim variable frequency to show}

\item{return.lim}{TRUE to return the xlim and ylim}

\item{...}{additional argument for plot}
}
\description{
Graphical illustration of selection uncertainty
}
\examples{
\dontrun{
X = matrix(rnorm(1000), nrow = 100)
Y = rowSums(X[,1:3])+rnorm(100)
mod.0 = csuv(X, Y, intercept = FALSE, q = 0, method.names = NULL)
cv.mod = lm.cv(X, Y, intercept = FALSE, fit.percent = 0.5, num.repeat = 50)
compare.mod = lm.compare.method(X, Y, intercept = FALSE)
plot(mod.0, compare.method.fit = compare.mod, cv.mod = cv.mod$est.b)
}
}
