% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csuv.r
\name{print.csuv}
\alias{print.csuv}
\title{Print the coefficients of csuv}
\usage{
\method{print}{csuv}(x, ...)
}
\arguments{
\item{x}{output of csuv()}

\item{...}{additional arguments to "print"}
}
\value{
return value from print(x$est.b)
}
\description{
Print the coefficients of csuv
}
