% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_helper.r
\name{csuv.ci}
\alias{csuv.ci}
\title{Graphical illustration of selection uncertainty}
\usage{
csuv.ci(csuv.fit, level, type = "original")
}
\arguments{
\item{csuv.fit}{fitted results from CSUV::csuv()}

\item{level}{significance level}

\item{type}{todo}
}
\description{
Graphical illustration of selection uncertainty
}
\examples{
\dontrun{
X = matrix(rnorm(1000), nrow = 100)
Y = rowSums(X[,1:3])+rnorm(100)
mod.0 = csuv(X, Y, intercept = FALSE, q = 0, method.names = NULL)
print(csuv.ci(mod.0, level = 0.1, type = "original"))
}
}
