% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{interactive.uncertainty.illustration}
\alias{interactive.uncertainty.illustration}
\title{Interactive version of the uncertainty illustration}
\usage{
interactive.uncertainty.illustration(log.level = NULL)
}
\arguments{
\item{log.level}{log level to set. Default is NULL, which means no change in log level. See the function CSUV::set.log.level for more details}
}
\description{
Interactive version of the uncertainty illustration
}
\examples{
\donttest{
interactive.uncertainty.illustration()
}
}
