% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.r
\name{set.log.level}
\alias{set.log.level}
\title{Set the level of logger}
\usage{
set.log.level(level)
}
\arguments{
\item{level}{log level, setting the level to futile.logger::DEBUG provides most details log, whereas setting the level to futile.logger::WARN provides least details log}
}
\value{
None
}
\description{
Set the level of logger
}
\examples{
\donttest{
set.log.level(futile.logger::DEBUG)
set.log.level(futile.logger::INFO)
set.log.level(futile.logger::WARN)
}
}
